Clean up, refinements, theme change.
It is spring and that means it is time to do something a little
different. After years of using solarized dark . . . I'm going to really
shake things up and use solarized light.

Watch out world.

This also condenses my config and hopefully makes it a little easier to
read / understand.

 Changes to be committed:
	modified:	README.org
	modified:	init-first.el
	modified:	init.el
	modified:	lisp/misc.el
	modified:	lisp/mode-specific-config.el